codigo/Live265/dt/pyproject.toml
[tool.poetry]
name = "dt"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
psycopg = {extras = ["binary"], version = "^3.1.18"}
celery = "^5.3.6"
minio = "^7.2.5"
opentelemetry-distro = "^0.45b0"
opentelemetry-exporter-otlp = "^1.24.0"
opentelemetry-instrumentation-celery = "^0.45b0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
